container_test.go file divided into separate files
